+++ CSSMS3/src/CSSMS3Display.h
@@ -133,16 +133,16 @@ public:
 	Pages GetCurrentPage();
 	void RefreshCurrentPage();
 	void RefreshPage(Pages page);
-	static void PrevPage(byte value);
-	static void NextPage(byte value);
+	static void PrevPage(int value);
+	static void NextPage(int value);
 
 	TFT_eSPI* GetTFT();
 	byte GetDisplayBrightness();
-	static void SetDisplayBrightness(byte brightness);
+	static void SetDisplayBrightness(int brightness);
 
 	void AddDebugTextLine(String newLine);
-	static void ReportHeapStatus(byte value);
-	static void ShowFontTableFixed(byte value);
+	static void ReportHeapStatus(int value);
+	static void ShowFontTableFixed(int value);
 	void ShowFontTable(int32_t xTL, int32_t yTL);
 
 };
